Move publish status change logic into a service, so that it can be called also from entity managemenet pages.
