+++ src/app/management/forms/crag-form/crag-form.component.scss
@@ -4,3 +4,14 @@
 .mat-hint a {
   text-decoration: none;
 }
+
+.publish-status-actions {
+  text-align: right;
+  white-space: nowrap;
+  button {
+    text-transform: uppercase;
+    &:last-child {
+      margin-left: 8px;
+    }
+  }
+}
