+++ src/app/management/pages/crag-sector-routes/crag-sector-routes.component.scss
@@ -44,3 +44,14 @@
     z-index: 1;
   }
 }
+
+.publish-status-actions {
+  text-align: right;
+  white-space: nowrap;
+  button {
+    text-transform: uppercase;
+    &:last-child {
+      margin-left: 8px;
+    }
+  }
+}
