+++ src/app/management/pages/crag/management-get-crag.query.graphql
@@ -26,5 +26,12 @@ query ManagementGetCrag($id: String!) {
     user {
       id
     }
+    created
+    sectors {
+      id
+      routes {
+        id
+      }
+    }
   }
 }
