+++ src/styles/global.scss
@@ -73,6 +73,9 @@ p.mt-16,
 .mb-16 {
   margin-bottom: 16px;
 }
+.mb-8 {
+  margin-bottom: 8px;
+}
 
 .mb-32 {
   margin-bottom: 32px;
